Added a `Map` for `isSubgroup_Group` functions.
i.e. added a map to be able to lookup `isSubGroup_Group` functions names given the type of the value to test and what you want to test for.

e.g. `isConsonant_Phoneme :: Phoneme -> Bool` would check whether a value of type `Phoneme` is a `Consonant` or not. It would be located at `("Phoneme", "consonant")` in the `Map`.