Finished `makeTheParser` + fixed argument issue.
Finally made a working version of `makeTheParser`, along with a version of it meant to run in the REPL: `testTheParser`.

Also, fixed an issue in the generated code: many functions were still using (`AT.Parser` `a`) instead of (`StateT` `StateType` `AT.Parser` `a`). In addition, the starting parsers (the ones with lots of options) were missing an argument in both the type and the declaration pattern.