Added `endOfWord` etc... functions.
`makeTheParser` now correctly generates and names the functions for checking whether the next character is part of a valid grapheme or not. These are used for graphemes that must come either at the end of a word or in the middle of the word. There's also a function for parsing non-graphemes to be used when parsing between words.