Started work to make the two parts compatible.
Started working to make the phoneme portion and the parser portion interact. At the moment, the values produced by the phoneme generator don't quite line up with the values required by the parser generator, so there's still more work to be done before the two can be properly tested together.