Wrote code to generate auto-reset functions.
There are now two new functions: "resetAutoStates" and "resetAndContinue". The first one just takes a State record and resets the auto-off states to off. The second one modifies the current state of a `(StateT StateType Parser)` computation and then runs another command.

The reason for the second function running the next command is that it's easier to turn the operation into a no-op when there are no auto-off states; it can just be defined as `id` instead of `return ()`.