Changed how "auto-off" state changes are handled.
Changed it by merging the two "State.modify'" steps into one step, reducing the required number of actions. Did this by creating a list of `ModifyStateX` actions that correspond to resetting the state changes, and adding them into the state change if there isn't a change for that state already.