Improved QuasiQuoter Parser
Now gives a warning if a field is missing a colon/equal sign. Previously, it would get to a field without a colon/equal sign, stop parsing completely, and fail silently. Now, it will succeed (if there's a space/tab between the field name and the next word), but raise a warning.

Might still run into problems if you have a field name immediately followed by a quotation mark, but otherwise it seems to work well.