Started work on adding lookahead for parsers.
Added the ability to "import" groups/aspects/traits for the parser file. Still need to implement actually using them to look ahead. Also, since the import check requires knowing info about the phoneme file, there's now a new version of `parseOrthographyFile` that takes info about the properties and phonemes as input. As a result, I'll have to change around code that calls `parseOrthographyFile` so that it can supply `parseOrthographyFileNew` with the correct data.